requirements.txt
altgraph==0.17.3
appdirs==1.4.4
asgiref==3.5.2
asn1crypto==1.5.1
bdist-mpkg==0.5.0
bonjour==0.0.3
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
configparser==5.3.0
contextlib2==21.6.0
contourpy==1.0.5
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.1
cycler==0.11.0
defusedxml==0.7.1
distlib==0.3.6
dj-database-url==1.0.0
Django==4.1
django-cors-headers==3.13.0
django-stubs==1.12.0
django-stubs-ext==0.5.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
enum34==1.1.10
filelock==3.8.0
fonttools==4.38.0
future==0.18.2
idna==3.4
importlib-metadata==5.0.0
importlib-resources==5.10.0
ipaddress==1.0.23
itypes==1.2.0
Jinja2==3.1.2
kiwisolver==1.4.4
macholib==1.16.2
MarkupSafe==2.1.1
matplotlib==3.6.1
modulegraph==0.19.3
mypy==0.982
mypy-extensions==0.4.3
numpy==1.23.4
oauthlib==3.2.2
packaging==21.3
pathlib2==2.3.7.post1
pdfkit==1.0.0
Pillow==9.2.0
platformdirs==2.5.2
psycopg2-binary==2.9.5
py2app==0.28.4
pycparser==2.21
PyJWT==2.5.0
pyOpenSSL==22.1.0
pyparsing==3.0.9
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.4
requests==2.28.1
requests-oauthlib==1.3.1
scandir==1.10.0
scipy==1.9.2
singledispatch==3.7.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
tomli==2.0.1
types-pytz==2022.5.0.0
types-PyYAML==6.0.12
typing==3.7.4.3
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.12
virtualenv==20.16.5
xattr==0.9.9
zipp==3.9.0
